move isDynamicPlaceholder and getDynamicPlaceholderPattern util functions to sitecore-jss/layout
